requirements.txt
asgiref==3.11.1
colorama==0.4.6
coverage==7.13.5
Django==6.0.4
djangorestframework==3.17.1
iniconfig==2.3.0
packaging==26.2
pillow==12.2.0
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
pytest-cov==7.1.0
pytest-django==4.12.0
python-dotenv==1.2.2
sqlparse==0.5.5
tzdata==2026.2
